backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.44
mysql-connector-python==8.3.0
mutagen==1.47.0
python-multipart==0.0.6
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
argon2-cffi
numpy
scikit-learn
pandas